Deacidify legacypasswds
