+++ hackage-server.cabal
@@ -332,7 +332,9 @@ library
       Distribution.Server.Features.TarIndexCache.State
       Distribution.Server.Features.LegacyRedirects
       Distribution.Server.Features.LegacyPasswds
+      Distribution.Server.Features.LegacyPasswds.Acid
       Distribution.Server.Features.LegacyPasswds.Auth
+      Distribution.Server.Features.LegacyPasswds.Backup
       Distribution.Server.Features.PackageContents
       Distribution.Server.Features.AdminFrontend
       Distribution.Server.Features.AdminLog
