+++ src/Distribution/Server/Features/LegacyPasswds.hs
@@ -1,13 +1,12 @@
-{-# LANGUAGE DeriveDataTypeable, TypeFamilies, TemplateHaskell,
+{-# LANGUAGE DeriveDataTypeable, TypeFamilies,
              RankNTypes, NamedFieldPuns, RecordWildCards,
              RecursiveDo, BangPatterns #-}
 module Distribution.Server.Features.LegacyPasswds (
     initLegacyPasswdsFeature,
     LegacyPasswdsFeature(..),
-    LegacyPasswdsTable,
-    lookupUserLegacyPasswd,
-    enumerateAllUserLegacyPasswd,
   ) where
+
+import qualified Distribution.Server.Features.LegacyPasswds.Acid as Acid
 
 import Prelude hiding (abs)
 
@@ -25,13 +24,8 @@ import qualified Distribution.Server.Users.Types as Users
 import qualified Distribution.Server.Users.Users as Users
 import qualified Distribution.Server.Framework.Auth as Auth
 
-import Data.IntMap (IntMap)
 import qualified Data.IntMap as IntMap
 import qualified Data.ByteString.Lazy.Char8 as LBS -- ASCII data only (password hashes)
-
-import Data.SafeCopy (base, deriveSafeCopy)
-import Control.Monad.Reader (ask)
-import Control.Monad.State (get, put)
 
 import Distribution.Text (display)
 import Data.Version
@@ -45,82 +39,26 @@ import Network.URI (URI(..), uriToString)
 data LegacyPasswdsFeature = LegacyPasswdsFeature {
     legacyPasswdsFeatureInterface :: HackageFeature,
 
-    queryLegacyPasswds       :: forall m. MonadIO m => m LegacyPasswdsTable,
+    queryLegacyPasswds       :: forall m. MonadIO m => m Acid.LegacyPasswdsTable,
     updateDeleteLegacyPasswd :: forall m. MonadIO m => UserId -> m Bool
 }
 
 instance IsHackageFeature LegacyPasswdsFeature where
   getFeatureInterface = legacyPasswdsFeatureInterface
 
--------------------------
--- Types of stored data
---
-
-newtype LegacyPasswdsTable = LegacyPasswdsTable (IntMap LegacyAuth.HtPasswdHash)
-  deriving (Eq, Show)
-
-emptyLegacyPasswdsTable :: LegacyPasswdsTable
-emptyLegacyPasswdsTable = LegacyPasswdsTable IntMap.empty
-
-lookupUserLegacyPasswd :: UserId -> LegacyPasswdsTable -> Maybe LegacyAuth.HtPasswdHash
-lookupUserLegacyPasswd (UserId uid) (LegacyPasswdsTable tbl) =
-    IntMap.lookup uid tbl
-
-enumerateAllUserLegacyPasswd :: LegacyPasswdsTable -> [UserId]
-enumerateAllUserLegacyPasswd (LegacyPasswdsTable tbl) =
-    map UserId (IntMap.keys tbl)
-
-$(deriveSafeCopy 0 'base ''LegacyPasswdsTable)
-
-instance MemSize LegacyPasswdsTable where
-    memSize (LegacyPasswdsTable a) = memSize1 a
-
-------------------------------
--- State queries and updates
---
-
-getLegacyPasswdsTable :: Query LegacyPasswdsTable LegacyPasswdsTable
-getLegacyPasswdsTable = ask
-
-replaceLegacyPasswdsTable :: LegacyPasswdsTable -> Update LegacyPasswdsTable ()
-replaceLegacyPasswdsTable = put
-
-setUserLegacyPasswd :: UserId -> LegacyAuth.HtPasswdHash -> Update LegacyPasswdsTable ()
-setUserLegacyPasswd (UserId uid) udetails = do
-    LegacyPasswdsTable tbl <- get
-    put $! LegacyPasswdsTable (IntMap.insert uid udetails tbl)
-
-deleteUserLegacyPasswd :: UserId -> Update LegacyPasswdsTable Bool
-deleteUserLegacyPasswd (UserId uid) = do
-    LegacyPasswdsTable tbl <- get
-    if IntMap.member uid tbl
-      then do put $! LegacyPasswdsTable (IntMap.delete uid tbl)
-              return True
-      else return False
-
-
-makeAcidic ''LegacyPasswdsTable [
-    --queries
-    'getLegacyPasswdsTable,
-    --updates
-    'replaceLegacyPasswdsTable,
-    'setUserLegacyPasswd,
-    'deleteUserLegacyPasswd
-  ]
-
 
 ---------------------
 -- State components
 --
 
-legacyPasswdsStateComponent :: FilePath -> IO (StateComponent AcidState LegacyPasswdsTable)
+legacyPasswdsStateComponent :: FilePath -> IO (StateComponent AcidState Acid.LegacyPasswdsTable)
 legacyPasswdsStateComponent stateDir = do
-  st <- openLocalStateFrom (stateDir </> "db" </> "LegacyPasswds") emptyLegacyPasswdsTable
+  st <- openLocalStateFrom (stateDir </> "db" </> "LegacyPasswds") Acid.emptyLegacyPasswdsTable
   return StateComponent {
       stateDesc    = "Support for upgrading accounts from htpasswd-style passwords"
     , stateHandle  = st
-    , getState     = query st GetLegacyPasswdsTable
-    , putState     = update st . ReplaceLegacyPasswdsTable
+    , getState     = query st Acid.GetLegacyPasswdsTable
+    , putState     = update st . Acid.ReplaceLegacyPasswdsTable
     , backupState  = \backuptype users ->
         [csvToBackup ["htpasswd.csv"] (legacyPasswdsToCSV backuptype users)]
     , restoreState = legacyPasswdsBackup
@@ -131,10 +69,10 @@ legacyPasswdsStateComponent stateDir = do
 -- Data backup and restore
 --
 
-legacyPasswdsBackup :: RestoreBackup LegacyPasswdsTable
+legacyPasswdsBackup :: RestoreBackup Acid.LegacyPasswdsTable
 legacyPasswdsBackup = updatePasswdsBackup []
 
-updatePasswdsBackup :: [(UserId, LegacyAuth.HtPasswdHash)] -> RestoreBackup LegacyPasswdsTable
+updatePasswdsBackup :: [(UserId, LegacyAuth.HtPasswdHash)] -> RestoreBackup Acid.LegacyPasswdsTable
 updatePasswdsBackup upasswds = RestoreBackup {
     restoreEntry = \entry -> case entry of
       BackupByteString ["htpasswd.csv"] bs -> do
@@ -147,7 +85,7 @@ updatePasswdsBackup upasswds = RestoreBackup {
   , restoreFinalize =
       let tbl =  IntMap.fromList [ (uid, htpasswd)
                                  | (UserId uid, htpasswd) <- upasswds ] in
-      return $! LegacyPasswdsTable tbl
+      return $! Acid.LegacyPasswdsTable tbl
   }
 
 importHtPasswds :: CSV -> Restore [(UserId, LegacyAuth.HtPasswdHash)]
@@ -160,8 +98,8 @@ importHtPasswds = mapM fromRecord . drop 2
 
     fromRecord x = fail $ "Error processing user details record: " ++ show x
 
-legacyPasswdsToCSV :: BackupType -> LegacyPasswdsTable -> CSV
-legacyPasswdsToCSV backuptype (LegacyPasswdsTable tbl)
+legacyPasswdsToCSV :: BackupType -> Acid.LegacyPasswdsTable -> CSV
+legacyPasswdsToCSV backuptype (Acid.LegacyPasswdsTable tbl)
     = ([showVersion version]:) $
       (headers:) $
 
@@ -198,7 +136,7 @@ initLegacyPasswdsFeature env@ServerEnv{serverStateDir, serverTemplatesDir,
     return feature
 
 legacyPasswdsFeature :: ServerEnv
-                     -> StateComponent AcidState LegacyPasswdsTable
+                     -> StateComponent AcidState Acid.LegacyPasswdsTable
                      -> Templates
                      -> UserFeature
                      -> LegacyPasswdsFeature
@@ -233,12 +171,12 @@ legacyPasswdsFeature env legacyPasswdsState templates UserFeature{..}
     -- Queries and updates
     --
 
-    queryLegacyPasswds :: MonadIO m => m LegacyPasswdsTable
-    queryLegacyPasswds = queryState legacyPasswdsState GetLegacyPasswdsTable
+    queryLegacyPasswds :: MonadIO m => m Acid.LegacyPasswdsTable
+    queryLegacyPasswds = queryState legacyPasswdsState Acid.GetLegacyPasswdsTable
 
     updateDeleteLegacyPasswd :: MonadIO m => UserId -> m Bool
     updateDeleteLegacyPasswd uid =
-      updateState legacyPasswdsState (DeleteUserLegacyPasswd uid)
+      updateState legacyPasswdsState (Acid.DeleteUserLegacyPasswd uid)
 
     -- Request handlers
     --
@@ -258,7 +196,7 @@ legacyPasswdsFeature env legacyPasswdsState templates UserFeature{..}
         passwdhash   <- expectTextPlain
         when (not $ validHtpasswd passwdhash) errBadHash
         let htpasswd = LegacyAuth.HtPasswdHash (LBS.unpack passwdhash)
-        updateState legacyPasswdsState $ SetUserLegacyPasswd uid htpasswd
+        updateState legacyPasswdsState $ Acid.SetUserLegacyPasswd uid htpasswd
         noContent $ toResponse ()
       where
         validHtpasswd str = LBS.length str == 13
@@ -270,7 +208,7 @@ legacyPasswdsFeature env legacyPasswdsState templates UserFeature{..}
     handleUserHtpasswdDelete dpath = do
         guardAuthorised_ [InGroup adminGroup]
         uid     <- lookupUserName =<< userNameInPath dpath
-        deleted <- updateState legacyPasswdsState (DeleteUserLegacyPasswd uid)
+        deleted <- updateState legacyPasswdsState (Acid.DeleteUserLegacyPasswd uid)
         when (not deleted) errNoHtpasswd
         noContent $ toResponse ()
       where
@@ -283,12 +221,12 @@ legacyPasswdsFeature env legacyPasswdsState templates UserFeature{..}
         (uid, uinfo, passwd) <- LegacyAuth.guardAuthenticated
                                   (RealmName "Old Hackage site")
                                   users
-                                  (flip lookupUserLegacyPasswd legacyPasswds)
+                                  (flip Acid.lookupUserLegacyPasswd legacyPasswds)
         when (userStatus uinfo /= Users.AccountDisabled Nothing) errHasAuth
         let auth = Users.UserAuth (Auth.newPasswdHash Auth.hackageRealm (userName uinfo) passwd)
         updateSetUserAuth uid auth
         updateSetUserEnabledStatus uid True
-        updateState legacyPasswdsState (DeleteUserLegacyPasswd uid)
+        updateState legacyPasswdsState (Acid.DeleteUserLegacyPasswd uid)
         template <- getTemplate templates "htpasswd-upgrade-success.html"
         ok $ toResponse $ template []
       where
@@ -312,7 +250,7 @@ legacyPasswdsFeature env legacyPasswdsState templates UserFeature{..}
     onAuthFail (Auth.UserStatusError uid
                   UserInfo { userStatus = AccountDisabled Nothing }) = do
         legacyPasswds <- queryLegacyPasswds
-        case lookupUserLegacyPasswd uid legacyPasswds of
+        case Acid.lookupUserLegacyPasswd uid legacyPasswds of
           Nothing -> return Nothing
           Just _  -> return (Just err)
       where
